waylandsink: Get the last buffer Weston has rendered back into waylandsink when going from PAUSED to READY
Weston keeps the buffer of last rendering while it is being displayed,
so this patch makes Weston release it by hiding the surface.

This implementation is backported from a part of the following commit
in the upstream.

"waylandsink: remove the buffer from the surface when going PAUSED -> READY"
commit ID: c4616a550a9c9a22eeb108e2e5b9753590951880