waylandsink: Avoid duplicate rendering of the first frame
The basesink class once passes the first frame through the preroll method.
After that, the same one as passed in the preroll method is received
through the render method when the state goes to the PLAYING.
This behavior could lead to a reference count inconsistency because
Weston notifies waylandsink of only releasing one of them.
Also there is the potential that the read access from Weston
to the buffer which has already returned to the buffer pool occurs.