waylandsink: Don't use data and size belonging to the GstWlMeta structure when using wl_kms
Those two member variables are used to unmap preallocated buffers that
are mapped with mmap() or check if they are mapped with kms_bo_map()
in the finalization of the buffer pool.
In the case of wl_kms, whether KMS BOs are mmaped or not can be distinguished
by checking if GstBuffers have dmabuf file descriptors with gst_is_dmabuf_memory(),
so the extra prameters to see if unmapping is needed are unnecessary.