waylandsink: Set up a wl_surface_frame callback to sync the server if not using wl_display_sync
This change is a preliminary implementation for processing display events
in a separated thread from the streaming thread.

This patch can change the wl_display file descriptor to being readable by
making the server dispatch a wl_surface_frame event withount any other sync
operations and the separated thread can wait for the file descriptor
to be readable with poll() etc. This processing is the advantage when
wl_display_sync() is not used to sync the server and event dispatching
moves to the separated thread.