waylandsink: Wait for releasing all the buffers by the server when the playback stops
This patch makes sure that all the buffers we posted to
the wayland server are released and the reference counts for
each of the buffers are completely dropped at the finalization of
the buffer pool.

The buffer management that makes use of the hash table API
is partially backported from 51a2c694ad8dc56c5c8fb4c2f11ab7982eb54023.